# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Atlas - Strategic Thinking Partner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ProjectStatus"
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/recent:
    get:
      operationId: listRecentProjects
      tags: [projects]
      summary: List recently opened projects
      parameters:
        - name: withinHours
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 720
            default: 48
      responses:
        "200":
          description: Recently opened projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentProjectsResponse"
        "400":
          description: Invalid query
  /projects/{projectId}/touch:
    post:
      operationId: touchProject
      tags: [projects]
      summary: Mark a project as recently opened
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Project touch recorded
        "404":
          description: Not found
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/summary:
    get:
      operationId: getProjectSummary
      tags: [projects]
      summary: Get project summary stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
  /projects/{id}/readiness-snapshots:
    get:
      operationId: listReadinessSnapshots
      tags: [projects]
      summary: List readiness score snapshots for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of readiness snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReadinessSnapshot"
    post:
      operationId: recordReadinessSnapshot
      tags: [projects]
      summary: Record a readiness score snapshot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReadinessSnapshotBody"
      responses:
        "201":
          description: Created snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessSnapshot"
        "400":
          description: Invalid payload
...
components:
  schemas:
...
    ProjectStatus:
      type: string
      enum: [shaping, committed, archived]
    ProjectSurfaceMode:
      type: string
      enum: [ambient, operational]
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/ProjectStatus"
        surfaceMode:
          $ref: "#/components/schemas/ProjectSurfaceMode"
        shape:
          type: object
          additionalProperties: true
          nullable: true
        workingTitle:
          type: ["string", "null"]
        committedAt:
          type: ["string", "null"]
          format: date-time
        entityType:
          type: string
          enum: [project, idea]
        memory:
          type: ["string", "null"]
        previewUrl:
          type: ["string", "null"]
        githubToken:
          type: ["string", "null"]
        linkedRepo:
          type: ["string", "null"]
        nodeState:
          type: object
          additionalProperties: true
          nullable: true
        pushHistory:
          type: array
          items: {}
          nullable: true
        lastHandoverAt:
          type: ["string", "null"]
        lastHandoverHash:
          type: ["string", "null"]
        lastOpenedAt:
          type: string
        latestSnapshotScore:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - status
        - surfaceMode
        - shape
        - workingTitle
        - committedAt
        - entityType
        - lastOpenedAt
        - createdAt
        - updatedAt
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        entity_type:
          type: string
          enum: [project, idea]
        status:
          $ref: "#/components/schemas/ProjectStatus"
          default: committed
        surfaceMode:
          $ref: "#/components/schemas/ProjectSurfaceMode"
        shape:
          type: object
          additionalProperties: true
        workingTitle:
          type: ["string", "null"]
      required:
        - name
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/ProjectStatus"
        surfaceMode:
          $ref: "#/components/schemas/ProjectSurfaceMode"
        shape:
          type: object
          additionalProperties: true
        workingTitle:
          type: ["string", "null"]
        memory:
          type: ["string", "null"]
        previewUrl:
          type: ["string", "null"]
        githubToken:
          type: ["string", "null"]
        linkedRepo:
          type: ["string", "null"]
        nodeState:
          type: object
          additionalProperties: true
          nullable: true
        pushHistory:
          type: array
          items: {}
          nullable: true
        lastHandoverAt:
          type: ["string", "null"]
        lastHandoverHash:
          type: ["string", "null"]
    RecentProject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          $ref: "#/components/schemas/ProjectStatus"
        last_opened_at:
          type: string
      required:
        - id
        - name
        - status
        - last_opened_at
    RecentProjectsResponse:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/RecentProject"
      required:
        - projects
    ProjectSummary:
      type: object
      properties:
        projectId:
          type: integer
        sessionCount:
          type: integer
        committedCount:
          type: integer
        parkedCount:
          type: integer
        recentMode:
          type: ["string", "null"]
      required:
        - projectId
        - sessionCount
        - committedCount
        - parkedCount
...
    ReadinessSnapshot:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        score:
          type: integer
        recordedAt:
          type: string
      required:
        - id
        - projectId
        - score
        - recordedAt
    CreateReadinessSnapshotBody:
      type: object
      properties:
        score:
          type: integer
          minimum: 0
          maximum: 100
      required:
        - score
...